look into millis:NaN.  Was treating colId as unique, but it is not.  Remake many getLoc queries. repair ghUtils.checkForPV2 error handling.  remove unused methods in ghV2.
test failure exposed a real issue in gh2du.PNP.  The test is wrong, and the attempt to bail in accr is wrong.
If fromCard, there is no longer a way to create in ACCR/reserved.   By hand, cards are born in No Status, then have to be moved.
In this case, split SHOULD occur, then be treated normally.